backport of commit 7b253d59fc23ac9265d7acf55ff0a74202307cba
