+++ testing/internal/e2e/tests/base/target_tcp_connect_test.go
@@ -5,7 +5,6 @@ package base_test
 
 import (
 	"context"
-	"fmt"
 	"strings"
 	"testing"
 
@@ -140,46 +139,4 @@ func TestCliTcpTargetConnectTargetViaTargetAndScopeNames(t *testing.T) {
 	hostIp = parts[len(parts)-1]
 	require.Equal(t, c.TargetAddress, hostIp, "SSH session did not return expected output")
 	t.Log("Successfully connected to target by its name and scope ID")
-}
-
-func TestCliTcpTargetConnectTargetWithTargetClientPort(t *testing.T) {
-	e2e.MaybeSkipTest(t)
-	c, err := loadTestConfig()
-	require.NoError(t, err)
-
-	ctx := context.Background()
-	boundary.AuthenticateAdminCli(t, ctx)
-	newOrgId := boundary.CreateNewOrgCli(t, ctx)
-	t.Cleanup(func() {
-		ctx := context.Background()
-		boundary.AuthenticateAdminCli(t, ctx)
-		output := e2e.RunCommand(ctx, "boundary", e2e.WithArgs("scopes", "delete", "-id", newOrgId))
-		require.NoError(t, output.Err, string(output.Stderr))
-	})
-
-	var expPort uint32 = 8356
-
-	newProjectId := boundary.CreateNewProjectCli(t, ctx, newOrgId)
-	newHostCatalogId := boundary.CreateNewHostCatalogCli(t, ctx, newProjectId)
-	newHostSetId := boundary.CreateNewHostSetCli(t, ctx, newHostCatalogId)
-	newHostId := boundary.CreateNewHostCli(t, ctx, newHostCatalogId, c.TargetAddress)
-	boundary.AddHostToHostSetCli(t, ctx, newHostSetId, newHostId)
-	newTargetId := boundary.CreateNewTargetCli(t, ctx, newProjectId, c.TargetPort, target.WithDefaultClientPort(expPort))
-	boundary.AddHostSourceToTargetCli(t, ctx, newTargetId, newHostSetId)
-
-	// Connect to target and print host's IP address
-	output := e2e.RunCommand(ctx, "boundary",
-		e2e.WithArgs(
-			"connect",
-			"-target-id", newTargetId,
-			"-exec", "/bin/echo", "--",
-			"{{boundary.port}}",
-		),
-	)
-	require.NoError(t, output.Err, string(output.Stderr))
-
-	parts := strings.Fields(string(output.Stdout))
-	port := parts[len(parts)-1]
-
-	require.Equal(t, fmt.Sprintf("%d", expPort), port, "specified port was not found")
 }
